+++ tests/test_scene.py
@@ -274,6 +274,17 @@ def test_scene_upload_sync(dataset_scene):
     assert response["dataset_id"] == dataset_scene.id
     assert response["new_scenes"] == len(scenes)
 
+    uploaded_scenes = dataset_scene.scenes
+    assert len(uploaded_scenes) == len(scenes)
+    assert all(
+        u["reference_id"] == o.reference_id
+        for u, o in zip(uploaded_scenes, scenes)
+    )
+    assert all(
+        u["metadata"] == o.metadata or (not u["metadata"] and not o.metadata)
+        for u, o in zip(uploaded_scenes, scenes)
+    )
+
 
 @pytest.mark.skip("Deactivated sync upload for scenes")
 @pytest.mark.integration
@@ -288,6 +299,17 @@ def test_scene_and_cuboid_upload_sync(dataset_scene):
 
     assert response["dataset_id"] == dataset_scene.id
     assert response["new_scenes"] == len(scenes)
+
+    uploaded_scenes = dataset_scene.scenes
+    assert len(uploaded_scenes) == len(scenes)
+    assert all(
+        u["reference_id"] == o.reference_id
+        for u, o in zip(uploaded_scenes, scenes)
+    )
+    assert all(
+        u["metadata"] == o.metadata or (not u["metadata"] and not o.metadata)
+        for u, o in zip(uploaded_scenes, scenes)
+    )
 
     lidar_item_ref = payload[SCENES_KEY][0][FRAMES_KEY][0]["lidar"][
         REFERENCE_ID_KEY
@@ -340,6 +362,17 @@ def test_scene_upload_async(dataset_scene):
         "total_steps": 1,
     }
 
+    uploaded_scenes = dataset_scene.scenes
+    assert len(uploaded_scenes) == len(scenes)
+    assert all(
+        u["reference_id"] == o.reference_id
+        for u, o in zip(uploaded_scenes, scenes)
+    )
+    assert all(
+        u["metadata"] == o.metadata or (not u["metadata"] and not o.metadata)
+        for u, o in zip(uploaded_scenes, scenes)
+    )
+
 
 @pytest.mark.integration
 def test_scene_upload_and_update(dataset_scene):
@@ -347,7 +380,6 @@ def test_scene_upload_and_update(dataset_scene):
     scenes = [
         LidarScene.from_json(scene_json) for scene_json in payload[SCENES_KEY]
     ]
-    reference_ids = [s.reference_id for s in scenes]
     update = payload[UPDATE_KEY]
 
     job = dataset_scene.append(scenes, update=update, asynchronous=True)
@@ -372,10 +404,16 @@ def test_scene_upload_and_update(dataset_scene):
         "total_steps": 1,
     }
 
-    fetched_scenes = [
-        dataset_scene.get_scene(ref_id) for ref_id in reference_ids
-    ]
-    assert len(fetched_scenes) == len(scenes)
+    uploaded_scenes = dataset_scene.scenes
+    assert len(uploaded_scenes) == len(scenes)
+    assert all(
+        u["reference_id"] == o.reference_id
+        for u, o in zip(uploaded_scenes, scenes)
+    )
+    assert all(
+        u["metadata"] == o.metadata or (not u["metadata"] and not o.metadata)
+        for u, o in zip(uploaded_scenes, scenes)
+    )
 
     job2 = dataset_scene.append(scenes, update=True, asynchronous=True)
     job2.sleep_until_complete()