+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to the [Nucleus Python Client](https://github.com/scaleapi/n
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.5.1) - 2021-01-11
+
+### Fixed
+- Updated dependencies for full Python 3.6 compatibility
+
+## [0.5.0](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.5.0) - 2021-01-10
+
+### Added
+- `nucleus.metrics` module for computing metrics between Nucleus `Annotation` and `Prediction` objects.
+
+## [0.4.5](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.4.5) - 2021-01-07
+
+### Added
+- `Dataset.scenes` property that fetches the Scale-generated ID, reference ID, type, and metadata of all scenes in the Dataset.
+
+## [0.4.4](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.4.4) - 2021-01-04
+
+### Added
+- `Slice.export_raw_items()` method that fetches accessible (signed) URLs for all items in the Slice.
+
+## [0.4.3](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.4.3) - 2022-01-03
+
+### Added
+- Improved error messages for categorization
+
+### Changed
+- Category taxonomies are now updatable
+
 ## [0.4.2](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.4.2) - 2021-12-16
 
 ### Added